ejemplo/requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
Requests==2.32.5
